requirements.txt
appdirs==1.4.4
black==20.8b1
click==7.1.2
IMAPClient==2.2.0
mypy-extensions==0.4.3
pathspec==0.8.1
python-dotenv==0.16.0
regex==2021.3.17
six==1.15.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
